+++ spec/lib/karafka/instrumentation/logger_listener_spec.rb
@@ -371,7 +371,7 @@ RSpec.describe_current do
 
     context 'when it is a connection.client.rebalance_callback.error' do
       let(:type) { 'connection.client.rebalance_callback.error' }
-      let(:message) { "Rebalance callack error occurred: #{error}" }
+      let(:message) { "Rebalance callback error occurred: #{error}" }
 
       it { expect(Karafka.logger).to have_received(:error).with(message) }
     end