+++ CHANGELOG.md
@@ -71,7 +71,7 @@
 2. Replace `Karafka::Pro::BaseConsumer` references to `Karafka::BaseConsumer`.
 3. Replace `Karafka::Instrumentation::Vendors::Datadog:Listener` with `Karafka::Instrumentation::Vendors::Datadog::MetricsListener`.
 
-## 2.0.41 (2023-14-19)
+## 2.0.41 (2023-04-19)
 - **[Feature]** Provide `Karafka::Pro::Iterator` for anonymous topic/partitions iterations and messages lookups (#1389 and #1427).
 - [Improvement] Optimize topic lookup for `read_topic` admin method usage.
 - [Improvement] Report via `LoggerListener` information about the partition on which a given job has started and finished.